Update and rename BasicCalc2 to BasicCalculator2.py

+++ BasicCalculator2.py
@@ -2,6 +2,6 @@
 
 num1 = input("Number 1")
 num2 = input("Number 2")
-result = float(num1) + float(num2)
+result = int(num1) + int(num2)
 
 print(result)
